add header files and reduce scope of local variable/functions
