+++ include/code_801DBB3C.h
@@ -1,0 +1,23 @@
+#ifndef POKEREVO_CODE_801DBB3C_H
+#define POKEREVO_CODE_801DBB3C_H
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+#include "code_801DB040.h"
+
+struct unkClass2
+{
+    u8 pad[0x24];
+};
+
+unkClass* func_801DBC58(unkClass2* p1);
+void func_801DBD00(unkClass2* p1, unkClass* p2);
+unkClass2* func_801DBD74(u32 p1, u32 p2);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif //POKEREVO_CODE_801DBB3C_H
